# OpInfo : index_fill (port remaining method_tests)