@mruberry 2021-04-27
> @mruberry addressed the review.
> 
> Is there a need for issue stating time for `index_fill` tests is high (around 90 seconds)?
> 
> Complex gradcheck themselves take around 50s. (details are in the description)
> 
> ```
> 40.88s call     test/test_ops.py::TestGradientsCUDA::test_fn_gradgrad_index_fill_cuda_complex128
> 13.12s call     test/test_ops.py::TestGradientsCPU::test_fn_gradgrad_index_fill_cpu_complex128
> 7.03s call     test/test_ops.py::TestGradientsCUDA::test_fn_grad_index_fill_cuda_complex128
> ```

I think @soulitzer is enabling a faster gradcheck soon, but would you file an issue for index_fill's test taking so long, too, @kshitij12345, just so we don't lose track of it?